% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allprecip_function.R
\name{allprecip}
\alias{allprecip}
\title{All Precipitation Function}
\usage{
allprecip(file_name)
}
\arguments{
\item{file_name}{}
}
\value{
RasterBrick of all precipitation values in a single file
}
\description{
All Precipitation Function
}
\examples{


\dontrun{

allprecip("./avancedr/spatial/data/raw/CZ_pr_EUR-11_CNRM-CERFACS-CNRM-CM5_historical_r1i1p1_CNRM-ALADIN63_v2_1hr_195101010030-195112312330.nc")

}


}
