% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf_function.R
\name{ddff}
\alias{ddff}
\title{Function for turning the file into a data.table showing precip depth in hourly intervals wit specific frequencies}
\usage{
ddff(dt, freq)
}
\arguments{
\item{dt}{data.table including the fields date (datum) and p (precipitation in kg/s)}

\item{freq}{the frequency to apply to the calculation}
}
\value{
data.table of precipitation depth split into durations on basis of hour (D1, D2, D3, D6, D12, D24) and specified frequencies
}
\description{
Function for turning the file into a data.table showing precip depth in hourly intervals wit specific frequencies
}
\examples{

\dontrun{

ddff(dta, frequency = c(2,5,10,20))

}

}
