% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{delta}
\alias{delta}
\title{Difference in historical baseline and rcp precipitation in climate models}
\format{
data.table with
\describe{
  \item{GCM}{Global Climate Model}
  \item{RCM}{Regional Climate Model}
  \item{RCP}{Which RCP of model, rcp 2.6, rcp 4.5, rcp 8.5}
  \item{Duration}{Division into hourly intervals, 1h, 2h, 3h, 6h, 12h, 24h}
  \item{freq}{Frequency of 2, 5, 10, 20}
  \item{hour}{hours, 1h, 2h, 3h, 6h, 12h, 24h}
  \item{delta}{historical depth divided by depth in mm/s}
  ...
}
}
\usage{
delta
}
\description{
Delta signifies the variable of the change
}
\keyword{datasets}
