% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{depth_maxima}
\alias{depth_maxima}
\title{Maximum depth of precipitation per climate model}
\format{
data.table with
\describe{
  \item{Duration}{Division into hourly intervals, 1h, 2h, 3h, 6h, 12h, 24h}
  \item{depth}{maximum precipitation in mm/s}
  \item{freq}{Frequency of 2, 5, 10, 20}
  \item{hour}{hours, 1h, 2h, 3h, 6h, 12h, 24h}
  \item{RCP}{Which RCP of model, rcp 2.6, rcp 4.5, rcp 8.5}
  \item{GCM}{Global Climate Model}
  \item{RCM}{Regional Climate Model}
  ...
}
}
\usage{
depth_maxima
}
\description{
Additionally separated according to GCM and RCM
}
\keyword{datasets}
