% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrecipDF.R
\name{getPrecipDF}
\alias{getPrecipDF}
\title{Get Precipitation Data Frame/ Data Table}
\usage{
getPrecipDF(list_of_files)
}
\arguments{
\item{list_of_files}{list of files including precipitation and spatial data}
}
\value{
data.table with date, id, precipitation, lon, lat and climate model columns
}
\description{
Get Precipitation Data Frame/ Data Table
}
\examples{

\dontrun{

getPrecipDF(list.files(path = "./avancedr/spatial/data/raw",
                       pattern = ".nc",
                       full.names = TRUE))

}

}
