% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prague_wgs}
\alias{prague_wgs}
\title{Spatial Polygon Data Frame representing Prague}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 1 rows and 8 columns.
}
\usage{
prague_wgs
}
\description{
Spatial Polygon Data Frame representing Prague
}
\keyword{datasets}
