% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{precip}
\alias{precip}
\title{Historical precipitation data sample}
\format{
data.table with
\describe{
  \item{Date}{Date Y-M-D}
  \item{Time}{Hourly data in 1h intervals}
  \item{Precipitation}{Precipitation in mm/s}
  \item{Longtitute}{Longtitude spatial data}
  \item{Latitude}{Latitude spatial data}
  \item{Climate Model}{Full name of climate model including GCM and RCM}
  ...
}
}
\usage{
precip
}
\description{
Precipitation data in hourly intervals deparated into climate models
}
\keyword{datasets}
