% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleprecip_function.R
\name{singleprecip}
\alias{singleprecip}
\title{Single Precipitation Plot Function}
\usage{
singleprecip(file_name, layer_numeric)
}
\arguments{
\item{file_name}{exact file from which to extract values}

\item{layer_numeric}{number representing which layer to plot}
}
\value{
plot of precipitation values with CZ spatial polygon inside
}
\description{
Single Precipitation Plot Function
}
\examples{

 \dontrun{

singleprecip("CZ_pr_EUR-11_CNRM-CERFACS-CNRM-CM5_historical_r1i1p1_CNRM-ALADIN63_v2_1hr_195101010030-195112312330.nc", 2)

}


}
